Rename ticker.py → price_ticker.py and test_ticker.py → test_price_ticker.py
Aligns the filename with the class name (PriceTicker) for clarity.
